% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_openadd.R
\name{as_openadd}
\alias{as_openadd}
\title{Coerce to openadd object}
\usage{
as_openadd(country = NULL, state = NULL, city = NULL, ...)
}
\arguments{
\item{country}{(characater) Country name}

\item{state}{(characater) State (or province) name}

\item{city}{(characater) City name}

\item{...}{ignored}
}
\description{
Coerce to openadd object
}
\examples{
\dontrun{
as_openadd(country="us", state="nv", city="las_vegas")
}
}

