% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/combine.R
\name{combine}
\alias{combine}
\title{Combine data helper}
\usage{
combine(...)
}
\arguments{
\item{...}{Data sets, all of class \code{oa}}
}
\description{
Combine data helper
}
\details{
Gives back a single \code{tbl_df}, with a subset of columns, currently
\code{lon}, \code{lat}, and \code{address}
}
\examples{
\dontrun{
dat <- list_data()
out1 <- oa_get(dat[5])
out2 <- oa_get(dat[32])
(alldat <- combine(out1, out2))
}
}

