% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/list_data.R
\name{list_data}
\alias{list_data}
\title{List available data}
\usage{
list_data(...)
}
\arguments{
\item{...}{Pass on curl options to \code{\link[httr]{GET}}}
}
\description{
List available data
}
\examples{
\dontrun{
res <- list_data()
res[5]
}
}

