% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oa_get.R
\name{oa_get}
\alias{oa_get}
\title{Get data from Openaddresses}
\usage{
oa_get(x, path = "~/.openadds", ...)
}
\arguments{
\item{x}{State name}

\item{path}{Path to store files in, a directory, not the file name}

\item{...}{Pass on curl options to \code{\link[httr]{GET}}}
}
\description{
Get data from Openaddresses
}
\examples{
\dontrun{
res <- oa_get(x = "20150316/us-nc-burke.csv")
res <- list_data()
(dat <- oa_get(res[5]))
(dat2 <- oa_get(res[17]))
}
}
\references{
\url{http://openaddresses.io/}
}

