% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_data.R
\name{oa_list}
\alias{oa_list}
\title{List available data}
\usage{
oa_list(...)
}
\arguments{
\item{...}{Pass on curl options to \code{\link[httr]{GET}}}
}
\description{
List available data
}
\examples{
\dontrun{
res <- oa_list()
res[5]
}
}

