% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oa_search.R
\name{oa_search}
\alias{oa_search}
\title{Search Openaddresses datasets}
\usage{
oa_search(country = NULL, state = NULL, city = NULL, ext = NULL, ...)
}
\arguments{
\item{country}{(characater) Country name}

\item{state}{(characater) State (or province) name}

\item{city}{(characater) City name}

\item{ext}{(characater) One of \code{.zip} or \code{.csv}}

\item{...}{Pass on curl options to \code{\link[httr]{GET}}}
}
\description{
Search Openaddresses datasets
}
\examples{
\dontrun{
# return all data in a data.frame
oa_search()

# search by various combinations
oa_search(country = "us")
oa_search(country = "us", state = "ca")
oa_search(state = "tx")
oa_search(city = "houston")
oa_search(ext = ".zip")
}
}

