% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openadds-package.R
\docType{package}
\name{openadds-package}
\alias{openadds-package}
\alias{openadds}
\title{R client for Openaddresses.io data}
\description{
R client for Openaddresses.io data
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}
