% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.openadd.R
\name{as_openadd}
\alias{as_openadd}
\title{Coerce to openadd object}
\usage{
as_openadd(country = NULL, state = NULL, city = NULL, ext = NULL, ...)
}
\arguments{
\item{...}{Ignored}

\item{x}{Input}
}
\description{
Coerce to openadd object
}
\examples{
\dontrun{
as_openadd("us", "nm", "hidalgo")
}
}

