% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_openadd.R
\name{as_openadd}
\alias{as_openadd}
\title{Coerce to openadd object}
\usage{
as_openadd(country = NULL, state = NULL, city = NULL, ...)
}
\arguments{
\item{country}{(characater) Country name}

\item{state}{(characater) State (or province) name}

\item{city}{(characater) City name}

\item{...}{ignored}
}
\value{
an object of class \code{openadd}
}
\description{
Coerce to openadd object
}
\details{
This is a helper function to let the user specify what they want
with any combination of country, state, and city - the output of which
can be passed to \code{\link{oa_get}} to get data.

If your search results in more than 1 result, we stop with message to
refine your search.
}
\examples{
\dontrun{
as_openadd(country="us", state="nv", city="las_vegas")

# too coarse, will ask you to refine your search
# as_openadd(country="us", state="mi", city="detroit")
}
}

