% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oa_get.R
\name{oa_get}
\alias{oa_get}
\title{Get data from Openaddresses}
\usage{
oa_get(x, path = "~/.openadds", ...)
}
\arguments{
\item{x}{URL for an openaddresses dataset, or an object of class openadd}

\item{path}{Path to store files in, a directory, not the file name}

\item{...}{Pass on curl options to \code{\link[httr]{GET}}}
}
\description{
Get data from Openaddresses
}
\examples{
\dontrun{
dat <- oa_list()
urls <- na.omit(dat$processed)
(out1 <- oa_get(urls[6]))
(out2 <- oa_get(urls[12]))
(out3 <- oa_get(urls[32]))
(out4 <- oa_get(urls[876]))
(out5 <- oa_get(urls[376]))
(out6 <- oa_get(urls[474]))
(out7 <- oa_get(urls[121]))
(out8 <- oa_get(urls[41]))
(out9 <- oa_get(urls[400]))

# from a openadd class object
oa_get(as_openadd(country="us", state="nv", city="las_vegas"))

# combine data sets
(alldat <- oa_combine(out1, out3))

# Map data
library("leaflet")
small <- out9$data[1:10000L, ]
leaflet(small) \%>\%
  addTiles() \%>\%
  addCircles(lat = ~LAT, lng = ~LON,
             popup = unname(apply(small[, c('NUMBER', 'STREET')], 1, paste, collapse = " ")))
}
}
\references{
\url{http://openaddresses.io/}
}

