% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_data.R
\name{oa_list}
\alias{oa_list}
\title{List available data}
\usage{
oa_list(...)
}
\arguments{
\item{...}{Pass on curl options to \code{\link[crul]{HttpClient}}}
}
\value{
A tibble (a data.frame)
}
\description{
List available data
}
\examples{
\dontrun{
(res <- oa_list())

# mean address count per state/country/etc
mean(res$`address count`, na.rm = TRUE)
}
}

