% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/openadds-package.R
\docType{package}
\name{openadds-package}
\alias{openadds}
\alias{openadds-package}
\title{R client for Openaddresses data}
\description{
R client for Openaddresses data
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

